Added a C version of the binary search implementation
After much headache I have managed to wrangle C into a solution to the Mask problem. It's based on the same concept of performing a binary search but just way faster. No guarantee that there aren't major memory leaks or something. Also the windows version has only been tested in wine so idk if it will work on actual hardware.